third_party/WebKit/LayoutTests/fast/multicol/vertical-lr/balancing/balance-short-trailing-empty-block.html

<!DOCTYPE html>
<script src="../../../../resources/js-test.js"></script>
<p>Both words below should be in the the right column.</p>
<div id="console"></div>
<div id="mc" style="-webkit-writing-mode:vertical-lr; -webkit-columns:2; columns:2; orphans:1; widows:1; -webkit-column-rule:1px solid; column-rule:1px solid; background:olive;">
    <div style="width:120px;"></div>
    <div style="line-height:80px;">second<br>column</div>
    <div style="width:30px;"></div>
</div>
<script>
    description("Tests if a block crossing a column boundary is recorded properly for balancing");
    shouldBe("mc.offsetWidth", "190");
</script>
